Refactor(WinningResult): 파일 위치 수정
WinningResult 클래스의 파일 위치에 오류가 있어 수정한다.
